Feat: Added Check Authorization in Table and Rooms

+++ src/controllers/room/roomsControllers.ts
@@ -2,6 +2,7 @@ import { Response, Request } from 'express';
 import { connection } from '../../database/connection';
 import crypto from 'crypto';
 import { RoomProps } from '../../types/roomControllersTypes';
+import checkAuthorization from '../../utils/checkAuthorization';
 
 export default {
     async index(req: Request, res: Response) {
@@ -11,12 +12,19 @@ export default {
     },
 
     async create(req: Request, res: Response) {
+        const { authorization } = req.headers;
         const valuesRoomInsert: RoomProps = {
             id: crypto.randomBytes(4).toString('hex'),
             title: req.body.title,
         };
 
         try {
+            const validateAuthorization = await checkAuthorization(authorization, 2);
+
+            if (!validateAuthorization.status) {
+                throw new Error(validateAuthorization.error);
+            };
+            
             await connection<RoomProps>('room').insert(valuesRoomInsert);
 
             return res.json({ status: `Room ${valuesRoomInsert.title} has been successfully added` });
@@ -51,8 +59,14 @@ export default {
     },
     async Delete(req: Request, res: Response) {
         const { id: idRoom } = req.params;
+        const { authorization } = req.headers;
 
         try {
+            const validateAuthorization = await checkAuthorization(authorization, 2);
+
+            if (!validateAuthorization.status) {
+                throw new Error(validateAuthorization.error);
+            };
             await connection<RoomProps>('room')
                 .where('id', idRoom)
                 .delete();
